Chore: isValid 인식문제로 활성화 로직 변경 및 코드 리팩토링
